# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Files Exchanger API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/link_tokens/types": {
            "get": {
                "tags": [
                    "LinkTokens"
                ],
                "summary": "Get all supported token types.",
                "operationId": "getSupportedLinkTokenTypes",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/header-accept-type"
                    },
                    {
                        "$ref": "#/components/parameters/header-authorization-token"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful get operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Authentication failed. Bearer token mismatch."
                    }
                },
                "security": [
                    {
                        "bearer_auth": []
                    }
                ]
            }
        },
        "/api/link_tokens/{link_token}": {
            "get": {
                "tags": [
                    "LinkTokens"
                ],
                "summary": "Get full information about link token by it's id.",
                "operationId": "getLinkTokenById",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/header-accept-type"
                    },
                    {
                        "$ref": "#/components/parameters/header-authorization-token"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful get.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/LinkTokenResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Authentication failed. Bearer token mismatch."
                    },
                    "403": {
                        "description": "Forbidden. You are not allowed to view this token."
                    }
                },
                "security": [
                    {
                        "bearer_auth": []
                    }
                ]
            }
        },
        "/api/users/{userId}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get single user by id",
                "description": "Returns user information",
                "operationId": "getUserById",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/header-accept-type"
                    },
                    {
                        "$ref": "#/components/parameters/header-authorization-token"
                    },
                    {
                        "$ref": "#/components/parameters/user-id-path-parameter"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful get operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User with specified id not found."
                    },
                    "401": {
                        "description": "Authentication failed. Bearer token mismatch."
                    }
                },
                "security": [
                    {
                        "bearer_auth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "LinkTokenResource": {
                "properties": {
                    "type": {
                        "type": "string",
                        "example": "link_tokens"
                    },
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "attributes": {
                        "properties": {
                            "token": {
                                "type": "string",
                                "example": "9194e805-0408-4184-8ad1-22744d1ffe17"
                            },
                            "type": {
                                "type": "string",
                                "enum": [
                                    "disposable",
                                    "unlimited"
                                ]
                            },
                            "created_at": {
                                "type": "string",
                                "format": "date-time"
                            },
                            "updated_at": {
                                "type": "string",
                                "format": "date-time"
                            }
                        },
                        "type": "object"
                    },
                    "links": {
                        "properties": {
                            "self": {
                                "type": "string",
                                "example": "http://localhost:8000/api/link_tokens/1"
                            },
                            "resource": {
                                "type": "string",
                                "example": "http://localhost:8000/api/guest/files/9194e805-0408-4184-8ad1-22744d1ffe17"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "UserResource": {
                "properties": {
                    "type": {
                        "type": "string",
                        "example": "users"
                    },
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "attributes": {
                        "properties": {
                            "name": {
                                "type": "string",
                                "example": "John"
                            },
                            "email": {
                                "type": "string",
                                "example": "hello-world@gmail.com"
                            },
                            "email_verified_at": {
                                "type": "string",
                                "format": "date-time",
                                "nullable": true
                            }
                        },
                        "type": "object"
                    },
                    "links": {
                        "properties": {
                            "self": {
                                "type": "string",
                                "example": "http://localhost:8000/api/users/1"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "parameters": {
            "header-accept-type": {
                "name": "Accept",
                "in": "header",
                "description": "Accept type",
                "required": true,
                "schema": {
                    "type": "string",
                    "example": "application/json"
                }
            },
            "header-authorization-token": {
                "name": "Authorization",
                "in": "header",
                "description": "Api authorization user token",
                "required": true,
                "schema": {
                    "type": "string",
                    "example": "Bearer 9194773b-3f24-42bb-93ca-654557dd303c"
                }
            },
...
            "user-id-path-parameter": {
                "name": "userId",
                "in": "path",
                "description": "User id",
                "required": true,
                "schema": {
                    "type": "integer",
                    "example": "1"
                }
            }
        },
...
        "securitySchemes": {
            "bearer_auth": {
                "type": "http",
                "name": "bearer_auth",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        []
    ]
}